requirements.txt
asn1crypto==0.22.0
attrs==17.2.0
Automat==0.6.0
beautifulsoup4==4.6.0
cffi==1.11.0
constantly==15.1.0
cryptography==2.0.3
cssselect==1.0.1
Django==1.11.5
django-background-tasks==1.1.11
django-braces==1.11.0
django-compat==1.0.14
django-crispy-forms==1.6.1
django-tables2==1.11.0
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
lxml==4.0.0
mysqlclient==1.3.12
parsel==1.2.0
pkg-resources==0.0.0
pyasn1==0.3.5
pyasn1-modules==0.1.4
pycparser==2.18
PyDispatcher==2.0.5
pyOpenSSL==17.3.0
pytz==2017.2
queuelib==1.4.2
service-identity==17.0.0
six==1.11.0
Twisted==17.5.0
w3lib==1.18.0
zope.interface==4.4.2
